DiscordBot/requirements.txt
discord.py==2.6.4
requests==2.32.5
display_tty==1.1.12
aiosqlite==0.21.0
ask-question==1.2.12
colorama==0.4.6
